refactor(server): 拆出 AddServerModal 表单区块组件与交互 hook
